Resolved Code Rabbit issues including rename folder
