+++ tests/codes/eurocode/pr_en_1992_1_1_2023/chapter_8_ultimate_limit_states/test_formula_8_41.py
@@ -4,7 +4,7 @@ from typing import ClassVar
 
 import pytest
 
-from blueprints.codes.eurocode.pr_en_1992_1_2023.chapter_8_ultimate_limit_states.formula_8_41 import Form8Dot41InclinationCompressionField
+from blueprints.codes.eurocode.pr_en_1992_1_1_2023.chapter_8_ultimate_limit_states.formula_8_41 import Form8Dot41InclinationCompressionField
 from blueprints.validations import LessOrEqualToZeroError
