+++ tests/codes/eurocode/pr_en_1992_1_1_2023/chapter_8_ultimate_limit_states/test_formula_8_42.py
@@ -4,7 +4,7 @@ from typing import ClassVar
 
 import pytest
 
-from blueprints.codes.eurocode.pr_en_1992_1_2023.chapter_8_ultimate_limit_states.formula_8_42 import (
+from blueprints.codes.eurocode.pr_en_1992_1_1_2023.chapter_8_ultimate_limit_states.formula_8_42 import (
     Form8Dot42ShearStressResistanceReinforcement,
 )
 from blueprints.validations import NegativeValueError
